checkpoint

+++ 2022/template.rs
@@ -6,7 +6,7 @@ use std::{
 use itertools::Itertools;
 use regex::Regex;
 
-type InputType = Vec<(String, i32)>;
+type InputType = Vec<(String, i128)>;
 
 #[aoc_generator(day8)]
 fn parse_input_day8(input: &str) -> InputType {
@@ -14,18 +14,18 @@ fn parse_input_day8(input: &str) -> InputType {
     // input.chars().collect()
 
     // Map to ints
-    // input.lines().map(|x| x.parse::<i32>().unwrap()).collect()
+    // input.lines().map(|x| x.parse::<i128>().unwrap()).collect()
 
     // Regex
     input.lines().map(|x| {}).collect()
 }
 
 #[aoc(day8, part1)]
-pub fn solve_part1(input: &InputType) -> i32 {
+pub fn solve_part1(input: &InputType) -> i128 {
     0
 }
 
 #[aoc(day8, part2)]
-pub fn solve_part2(input: &InputType) -> i32 {
+pub fn solve_part2(input: &InputType) -> i128 {
     0
 }
